Create the Code
Completed Assignment: The code imports xml and sqlite3. Then it creates the database file ‘trackdb.sqlite’ and makes a database connection. It makes fresh tables using executescript(), deleting specific tables if they already exist in the database.

Next, the code creates a new “Artist” table with columns for id and name. The id/primary key will be created and auto-increment as it finds unique artists. The “Genre” table and “Album” table also go through these steps (adding the artist_id to “Album” as well). Lastly, it creates a new “Track” table with columns for (track) id, title, album_id, genre_id, length, rating, and count.

Once the tables are created, the code prompts the user for a file name, defaulting to ‘Library.xml’ if left blank. It performs a dictionary lookup, loops through the children in the outer dictionary, and finds the child tag that has a particular key.

Then, it parses the file, and finds and iterates through all of the third level dictionaries, skipping/continuing if there is no track id. It looks up Name, Artist, Album, Genre, Play Count, Rating, and Total time (skipping/continuing if anything is missing). And prints all the variables (name, artist, album, count, rating, length). 

The programs puts in new rows into the “Artist” table when it finds a new/unique artist, puts in the artist variable and gives the id. These same steps are then also done for the “Genre” and “Album” tables (and adds the artist_id to “Album”). Then the program puts in the track “title” variable, album_id, genre_id, length, rating, and count into the “Track” table. Finally, it commits (writes to memory).

Once run, the database can be opened, and the SQL query can be used to retrieve the data from the database.